minor api
